+++ controllers/diary.controller.js
