+++ routes/auth/index.js
@@ -1,10 +1,10 @@
 const express = require('express');
 const router = express.Router();
-const controller = require('../../controllers/userController');
+const controller = require('../../controllers/auth.controller');
 
-router.get('/auth/silent-refresh', controller.silentRefresh);
-router.post('/auth/login', controller.login);
-router.post('/auth/logout', controller.logout);
-router.post('/auth/register', controller.register);
+router.get('/silent-refresh', controller.silentRefresh);
+router.post('/login', controller.login);
+router.post('/logout', controller.logout);
+router.post('/register', controller.register);
 
 module.exports = router;
